requirements.txt
numpy>=1.26
scipy>=1.12
h5py>=3.10
pyyaml>=6.0
pytest>=8.0
torch>=2.2
